functions and methods

+++ R/utilities.R
@@ -113,21 +113,20 @@ runCommandline <- function(args, runid="01", make_bam=FALSE, del_sam=TRUE, dir=T
       logdir <- normalizePath(yamlinput(args)$results_path$path)
       dir.name <- dir.name
     }
-    ## Check if expected files exists or not
+    ## Check if expected files exists or not 
+    ## Important, this happen in the new location in the case of dir=TRUE
     return <- .checkOutArgs2(args, make_bam=make_bam, dir=dir, dir.name=dir.name)
     args.return <- return$args
     completed <- return$completed
-    ## Check if one sample/commandline expects one or more output files
-    outputList <- unlist((output(args)))
-    names(outputList) <- rep(names(output(args)), each=sum(lengths(output(args)[[1]])))
+    # ## Check if one sample/commandline expects one or more output files
+    # outputList <- unlist((output(args)))
+    # names(outputList) <- rep(names(output(args)), each=sum(lengths(output(args)[[1]])))
     ## Create log files
-    #file_cmdlist <- file.path(logdir, paste0("submitargs", runid, "_", dir.name, "_cmd_", format(Sys.time(), "%b%d%Y_%H%Ms%S")))
     file_log <- file.path(logdir, paste0("submitargs", runid, "_", dir.name, "_log_", format(Sys.time(), "%b%d%Y_%H%Ms%S")))
-   # file_err <- file.path(logdir, paste0("submitargs", runid, "_", dir.name, "_err_", format(Sys.time(), "%b%d%Y_%H%Ms%S")))
+    ## Sample status and Time 
     sample_status <- sapply(names(cmdlist(args)), function(x) list(NULL))
     time_status <- data.frame(Targets=names(cmdlist(args)), time_start=NA, time_end=NA)
     ## Progress bar
-    #cat("\n", crayon::blue("---- Running `cmdlist` ----"), "\n")
     pb <- txtProgressBar(min = 0, max = length(cmdlist(args)), style = 3)
     ## Check input
     if(length(args$inputvars) >= 1){
@@ -204,7 +203,7 @@ runCommandline <- function(args, runid="01", make_bam=FALSE, del_sam=TRUE, dir=T
     ## Status and log.files
     df.status <- data.frame(matrix(do.call("c", sample_status), nrow=length(sample_status), byrow=TRUE))
     colnames(df.status) <- files(args.return)$steps
-    check <- check.output(args.return)
+    check <- check.output(args)
     df.status.f <- cbind(check, df.status)
     df.status.f[c(2:4)] <- sapply(df.status.f[c(2:4)],as.numeric)
     ## time
@@ -215,6 +214,7 @@ runCommandline <- function(args, runid="01", make_bam=FALSE, del_sam=TRUE, dir=T
     args.return[["status"]]$status.completed <- df.status.f
     args.return[["status"]]$status.time <- time_status
     args.return[["files"]][["log"]] <- file_log
+    print(dir)
     ## Create recursive the subfolders
     if(dir==TRUE){
       for(i in seq_along(names(cmdlist(args)))){
@@ -231,16 +231,16 @@ runCommandline <- function(args, runid="01", make_bam=FALSE, del_sam=TRUE, dir=T
       args.return[["files"]][["log"]] <- file.path(logdir, dir.name, "_logs", basename(file_log))
       ## output FILES
       if(make_bam==TRUE) args.return <- .checkOutArgs2(args, make_bam=make_bam, dir=FALSE, dir.name=dir.name)$args
-      outputList_new <- as.character()
-      for(i in seq_along(output(args.return))){
-        if(length(output(args.return)[[i]]) > 1){
-          for(j in seq_along(output(args.return)[[i]])){
-            for(k in seq_along(output(args.return)[[i]][[j]])){
-              if(file.exists(output(args.return)[[i]][[j]])){
-                name <- strsplit(output(args.return)[[i]][[j]][[k]], split="\\/")[[1]]
+      #outputList_new <- as.character()
+      for(i in seq_along(output(args))){
+        if(length(output(args)[[i]]) > 1){
+            for(j in seq_along(output(args)[[i]])){
+            for(k in seq_along(output(args)[[i]][[j]])){
+              if(file.exists(output(args)[[i]][[j]])){
+                name <- strsplit(output(args)[[i]][[j]][[k]], split="\\/")[[1]]
                 name <- name[length(name)]
-                file.rename(from=output(args.return)[[i]][[j]][[k]], to=file.path(logdir, dir.name, name))
-                outputList_new <- c(outputList_new, file.path(logdir, dir.name, name))
+                file.rename(from=output(args)[[i]][[j]][[k]], to=file.path(logdir, dir.name, name))
+                #outputList_new <- c(outputList_new, file.path(logdir, dir.name, name))
                 # file.rename(from=output(args.return)[[i]][[j]][[k]], to=file.path(logdir, dir.name, names(output(args.return)[i]), name))
                 # outputList_new <- c(outputList_new, file.path(logdir, dir.name, names(output(args.return)[i]), name))
               } else if(!file.exists(output(args.return)[[i]][[j]][[k]])){
@@ -248,22 +248,22 @@ runCommandline <- function(args, runid="01", make_bam=FALSE, del_sam=TRUE, dir=T
               }
             }
           }
-        } else if(length(output(args.return)[[i]]) == 1){
-          for(j in seq_along(output(args.return)[[i]][[1]])){
-            if(file.exists(output(args.return)[[i]][[1]][[j]])){
-              name <- strsplit(output(args.return)[[i]][[1]][[j]], split="\\/")[[1]]
+        } else if(length(output(args)[[i]]) == 1){
+          for(j in seq_along(output(args)[[i]][[1]])){
+            if(file.exists(output(args)[[i]][[1]][[j]])){
+              name <- strsplit(output(args)[[i]][[1]][[j]], split="\\/")[[1]]
               name <- name[length(name)]
-              file.rename(from=output(args.return)[[i]][[1]][[j]], to=file.path(logdir, dir.name, name))
-              outputList_new <- c(outputList_new, file.path(logdir, dir.name, name))
+              file.rename(from=output(args)[[i]][[1]][[j]], to=file.path(logdir, dir.name, name))
+              # outputList_new <- c(outputList_new, file.path(logdir, dir.name, name))
               #file.rename(from=output(args.return)[[i]][[1]][[j]], to=file.path(logdir, dir.name, names(output(args.return)[i]), name))
               # outputList_new <- c(outputList_new, file.path(logdir, dir.name, names(output(args.return)[i]), name))
-            } else if(!file.exists(output(args.return)[[i]][[1]][[j]])){
+            } else if(!file.exists(output(args)[[i]][[1]][[j]])){
               dump <- "No such file or directory"
             }
           }
         }
       }
-      args.return <- output_update(args.return, dir=TRUE, dir.name=dir.name, replace=FALSE)
+      #args.return <- output_update(args.return, dir=TRUE, dir.name=dir.name, replace=FALSE)
     }
     ## double check output file
     cat("\n")
@@ -307,7 +307,7 @@ runCommandline <- function(args, runid="01", make_bam=FALSE, del_sam=TRUE, dir=T
 # .tryRunC("ls", "la")
 
 # runid="01"
-# make_bam=FALSE
+# make_bam=TRUE
 # del_sam=TRUE
 # dir=TRUE
 # dir.name=NULL
@@ -350,12 +350,14 @@ runCommandline <- function(args, runid="01", make_bam=FALSE, del_sam=TRUE, dir=T
     completed.bam <- grepl(".bam$", output_args)
     if(any(sam_files)){
       for(k in which(sam_files)){
-        Rsamtools::asBam(file=output_args[k], destination=gsub("\\.sam$", "", output_args[k]), overwrite=TRUE, indexDestination=TRUE)
-        if(del_sam==TRUE){
-          unlink(output_args[k])
-        } else if(del_sam==FALSE){
-          dump <- "do nothing"
-        }
+        if(file.exists(output_args[k])){
+          Rsamtools::asBam(file=output_args[k], destination=gsub("\\.sam$", "", output_args[k]), overwrite=TRUE, indexDestination=TRUE)
+          if(del_sam==TRUE){
+            unlink(output_args[k])
+          } else if(del_sam==FALSE){
+            dump <- "do nothing"
+          }
+        } else next ()
       } } else if(any(others_files)){
         dump <- "do nothing"
       }
@@ -377,21 +379,24 @@ runCommandline <- function(args, runid="01", make_bam=FALSE, del_sam=TRUE, dir=T
 ## if the expectedoutput has been created  ##
 ########################################################
 .checkOutArgs2 <- function(args, make_bam, dir=dir, dir.name){
-  print(dir)
-  suppressWarnings({
-    if(dir==TRUE){
-      if(!is.null(dir.name)){
-        if(!file.exists(normalizePath(file.path(yamlinput(args)$results_path$path, dir.name)))) {
-         # dir.create(normalizePath(file.path(yamlinput(args)$results_path$path, dir.name)), recursive = TRUE)
-        }
-        
-      } 
-    }
-  })
+  # print(dir)
+  # suppressWarnings({
+  #   if(dir==TRUE){
+  #     if(!is.null(dir.name)){
+  #       if(!file.exists(normalizePath(file.path(yamlinput(args)$results_path$path, dir.name)))) {
+  #        # dir.create(normalizePath(file.path(yamlinput(args)$results_path$path, dir.name)), recursive = TRUE)
+  #       }
+  #       
+  #     } 
+  #   }
+  # })
   if(make_bam==TRUE) {
     ## Validation for Hisat2
     if(any(grepl("samtools", names(clt(args))))){ stop("argument 'make_bam' should be 'FALSE' when using the workflow with 'SAMtools'")}
     args <- output_update(args, dir=dir, dir.name=dir.name, replace=TRUE, extension=c(".sam", ".bam"), make_bam=make_bam)
+  }
+  if(dir==TRUE) {
+    args <- output_update(args, dir=dir, dir.name=dir.name)
   }
   completed <- output(args)
   for(i in seq_along(output(args))){
